+++ mmtune/__main__.py
@@ -1,0 +1,5 @@
+"""The mmtune entry point."""
+from .run import main
+
+if __name__ == '__main__':
+    main()
